+++ Plugins/Extensions/EpgToXml/paths.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 
 CONFIG_DIR = "/media/hdd/epgtoxml"
 TASKS_PATH = CONFIG_DIR + "/tasks.json"
-OUTPUT_DIR = CONFIG_DIR + "/output"
+OUTPUT_DIR = "/tmp/epgtoxml/output"
 SETTINGS_PATH = CONFIG_DIR + "/settings.json"
 DEBUG_LOG_PATH = CONFIG_DIR + "/epgtoxml-debug.log"
 DEBUG_LOG_ROTATED_PATH = CONFIG_DIR + "/epgtoxml-debug.log.1"